# backend/src/main/resources/api/openapi.yml
openapi: 3.0.3
info:
  title: POS SaaS API
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local
  - url: https://api.pos-saas.com
    description: Production
paths:
...
  /api/v1/tenant/register:
    post:
      summary: Registrar tenant desde onboarding mobile
      operationId: registerTenant
      tags:
        - Tenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterTenantRequest'
            example:
              businessName: "Pollería Don Juan"
              phone: "+591 70123456"
              currency: "BOB"
              businessType: "restaurant"
              features:
                sizeVariants: false
                modifiers: true
                combos: true
                stockControl: false
                ingredientRecipes: false
                customerDisplay: false
                ticketPrinting: false
              adminPin: "1234"
      responses:
        '201':
          description: Tenant registrado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterTenantResponse'
        '400':
          description: Request inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: El teléfono ya está registrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/tenant/{tenantId}/plan:
    get:
      summary: Obtener plan activo y features del tenant
      operationId: getTenantPlan
      tags:
        - Tenant
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Plan activo y features habilitadas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanFeaturesResponse'
              example:
                planName: "basic"
                active: true
                endsAt: "2026-05-06T00:00:00Z"
                sync: true
                advancedReports: true
                maxCashiers: 2
                maxBranches: 1
                multiBranch: false
                telegramStockAlerts: true
                csvExport: true
...
  /api/v1/admin/tenants:
    get:
      summary: Listar tenants (panel admin)
      operationId: listTenants
      tags:
        - Admin
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Lista paginada de tenants
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedTenantSummaryResponse'
  /api/v1/admin/tenants/{id}:
    get:
      summary: Detalle de tenant
      operationId: getTenant
      tags:
        - Admin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Detalle completo del tenant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantDetailResponse'
        '404':
          description: Tenant no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    RegisterTenantRequest:
      type: object
      required:
        - businessName
        - phone
        - currency
        - businessType
        - features
        - adminPin
      properties:
        businessName:
          type: string
          minLength: 2
          maxLength: 60
        phone:
          type: string
          minLength: 7
          maxLength: 15
        currency:
          type: string
          enum: [BOB, USD]
        businessType:
          type: string
          enum: [restaurant, pizzeria, bakery, retail, other]
        features:
          $ref: '#/components/schemas/FeatureFlags'
        adminPin:
          type: string
          pattern: '^\d{4}$'
    FeatureFlags:
      type: object
      required:
        - sizeVariants
        - modifiers
        - combos
        - stockControl
        - ingredientRecipes
        - customerDisplay
        - ticketPrinting
      properties:
        sizeVariants:
          type: boolean
        modifiers:
          type: boolean
        combos:
          type: boolean
        stockControl:
          type: boolean
        ingredientRecipes:
          type: boolean
        customerDisplay:
          type: boolean
        ticketPrinting:
          type: boolean
    RegisterTenantResponse:
      type: object
      properties:
        tenantId:
          type: string
          format: uuid
        businessName:
          type: string
        plan:
          type: string
        branchId:
          type: string
          format: uuid
    TenantSummaryResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        phone:
          type: string
        planName:
          type: string
        createdAt:
          type: string
          format: date-time
    PagedTenantSummaryResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/TenantSummaryResponse'
        total:
          type: integer
          format: int64
        page:
          type: integer
        size:
          type: integer
    FeatureFlagsResponse:
      type: object
      properties:
        sizeVariants:
          type: boolean
        modifiers:
          type: boolean
        combos:
          type: boolean
        stockControl:
          type: boolean
        ingredientRecipes:
          type: boolean
        customerDisplay:
          type: boolean
        ticketPrinting:
          type: boolean
    TenantDetailResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        phone:
          type: string
        currency:
          type: string
        businessType:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        featuresConfig:
          $ref: '#/components/schemas/FeatureFlagsResponse'
        planId:
          type: string
          format: uuid
        planName:
          type: string
        subscriptionStatus:
          type: string
        subscriptionStartedAt:
          type: string
          format: date-time
        subscriptionEndsAt:
          type: string
          format: date-time
        subscriptionHistory:
          type: array
          items:
            $ref: '#/components/schemas/SubscriptionHistoryItem'
        branchId:
          type: string
          format: uuid
        branchName:
          type: string
    SubscriptionHistoryItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        planId:
          type: string
          format: uuid
        status:
          type: string
        startedAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        activatedAt:
          type: string
          format: date-time
...
    PlanFeaturesResponse:
      type: object
      properties:
        planName:
          type: string
          enum: [local, basic, pro]
        active:
          type: boolean
        endsAt:
          type: string
          format: date-time
          nullable: true
        sync:
          type: boolean
        advancedReports:
          type: boolean
        maxCashiers:
          type: integer
        maxBranches:
          type: integer
          nullable: true
        multiBranch:
          type: boolean
        telegramStockAlerts:
          type: boolean
        csvExport:
          type: boolean
    ErrorResponse:
      type: object
      properties:
        error:
          type: string